Add pipeline template variables
Removed one of the `resources` elements, as there should only be one.
Removed the `pr.branches.include`, as the `trigger.branches.include`
containing `main` means that it triggers on PRs as well.
Since the react-complete-build template contains jobs for "prod
tagging" as well as building, I removed the stuff for prod-tagging in
our pipeline.
Using the `$(Build.xyz)` variables directly in the values passed as
parameters instead of using our own variables defined in the pipeline
(the "Hemmeligheter" var group). I *think* this should work. Let's give
it a try. :)

Last, but not least, included a `coverage` script in `package.json`
because the react-complete-build template does a
[`yarn coverage`](https://github.com/statisticsnorway/azure-pipelines-templates/blob/d0d954de072cf79983e196a909989ef623ad4ecd/javascript/react-complete-build.yml#L25)
and will probably fail if no such script is found.
(The script content is directly copied from the dapla-react-ref-app.)

Internal ref [STRATUS-731]