Add coverage config + remove extra pipeline file
Copied the Jest coverage config from dapla-react-reference-app,
except for the coverage threshold config. We'll add the thresholds
when we figure out how we want the coverage (and tests) to work.
The coverage config is neccessary for the pipeline to work,
the PublishCodeCoverageResults task needs the files created by
cobertura (which is the coverage reporter added in the Jest
config).

Also removed the extra pipeline file I added by mistake earlier.

Internal ref [STRATUS-731]