Create our own pipeline
The pipeline does "everything"* when a PR is created -
except for the SonarQube analysis, since that can only
be done on `master` branch (due to restrictions in our
version of SonarQube).

* Run tests & check coverage, build the app, build a
Docker image & tag it as "unscanned", runs the vuln-scan
task and tags the image as "scanned" if the scan task is
OK.

There is a separate job that handles the SonarQube
analysis, which is triggered by changes on `master`
branch, *except* for PRs. We might have to tweak this
a bit, as there might be cases I haven't thought about
where there are changes to the `master` branch that are
not PRs (and hence will trigger the SQ job even if it's
unneccessary or unwanted).

Internal ref. [STRATUS-731]