Add tests to mock backend
To ready the code (and tests) for integration with the backend,
we needed an integration library and mocking for this.

We use the Axios-hooks library in HRForm, but there is no communication with
the backend, which means that the submit action will fail until an integration
is implemented.

Internal ref: [STRATUS-751]

Co-authored-by: larsroger <lrb@ssb.no>
Co-authored-by: Lisa <35797988+lisawolderiksen@users.noreply.github.com>
Co-authored-by: are3k <egk@ssb.no>