Add tooltips and improve formatting
Why: The form didn't react as wanted, some aliging was necessary, and
we needed to include tooltip popups to give extended information about
some of the fields in form.

How: We added some more elements from Semantic-ui-react and created a
data structure for helptexts that can be shown in popup tooltips.

int.ref: [STRATUS-748]

Co-authored-by: ArielKarlsen <54323769+ArielKarlsen@users.noreply.github.com>
Co-authored-by: ssbwep <36296800+ssbwep@users.noreply.github.com>