Bug fix image repo & improved helptext
The image repository field in bip-start requires 'http' prefix. This is
not valid when the helmrelease is deployed to the application clusters.
Therefore urls without https should be allowed.
As for now the validation of the image repository field is removed, but
could be reintroduced with a custom regular expression to reflect how
we want the url to look like.

Additionally, the helptexts for the fields have been improved and
more texts have been added.

This PR also includes some syntax fixes.